make base intermediate view
